+++ paimon-spark/paimon-spark-4.1/src/main/scala/org/apache/paimon/spark/catalyst/optimizer/MergePaimonScalarSubqueries.scala
@@ -90,4 +90,3 @@ object MergePaimonScalarSubqueries extends MergePaimonScalarSubqueriesBase {
     ScalarSubquery(plan, exprId = exprId)
   }
 }
-
