+++ paimon-spark/paimon-spark3-common/src/main/scala/org/apache/spark/sql/paimon/shims/Spark3Shim.scala
@@ -18,27 +18,32 @@
 
 package org.apache.spark.sql.paimon.shims
 
+import org.apache.hadoop.fs.Path
 import org.apache.paimon.data.variant.Variant
 import org.apache.paimon.spark.catalyst.analysis.Spark3ResolutionRules
 import org.apache.paimon.spark.catalyst.parser.extensions.PaimonSpark3SqlExtensionsParser
 import org.apache.paimon.spark.data.{Spark3ArrayData, Spark3InternalRow, Spark3InternalRowWithBlob, SparkArrayData, SparkInternalRow}
 import org.apache.paimon.types.{DataType, RowType}
-
 import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions.{Attribute, Expression}
 import org.apache.spark.sql.catalyst.expressions.aggregate.AggregateExpression
 import org.apache.spark.sql.catalyst.parser.ParserInterface
+import org.apache.spark.sql.catalyst.plans.logical.MergeRows.Instruction
 import org.apache.spark.sql.catalyst.plans.logical._
 import org.apache.spark.sql.catalyst.rules.Rule
 import org.apache.spark.sql.catalyst.util.ArrayData
 import org.apache.spark.sql.connector.catalog.{Identifier, Table, TableCatalog}
 import org.apache.spark.sql.connector.expressions.Transform
 import org.apache.spark.sql.execution.datasources._
+import org.apache.spark.sql.execution.streaming.runtime.MetadataLogFileIndex
+import org.apache.spark.sql.execution.streaming.sinks.FileStreamSink
 import org.apache.spark.sql.types.StructType
 import org.apache.spark.sql.util.CaseInsensitiveStringMap
 
 import java.util.{Map => JMap}
+
+import scala.collection.JavaConverters._
 
 class Spark3Shim extends SparkShim {
 
@@ -110,6 +115,13 @@ class Spark3Shim extends SparkShim {
       notMatchedBySourceActions)
   }
 
+  override def createKeep(
+      context: String,
+      condition: Expression,
+      output: Seq[Expression]): Instruction = {
+    MergeRows.Keep(condition, output)
+  }
+
   override def toPaimonVariant(o: Object): Variant = throw new UnsupportedOperationException()
 
   override def isSparkVariantType(dataType: org.apache.spark.sql.types.DataType): Boolean = false
@@ -123,17 +135,69 @@ class Spark3Shim extends SparkShim {
   override def toPaimonVariant(array: ArrayData, pos: Int): Variant =
     throw new UnsupportedOperationException()
 
-  override def createFileIndex(
-      options: CaseInsensitiveStringMap,
-      sparkSession: SparkSession,
-      paths: Seq[String],
-      userSpecifiedSchema: Option[StructType],
-      partitionSchema: StructType): PartitioningAwareFileIndex = {
-    MinorVersionShim.createFileIndex(
-      options,
-      sparkSession,
-      paths,
-      userSpecifiedSchema,
-      partitionSchema)
+  def createFileIndex(
+                       options: CaseInsensitiveStringMap,
+                       sparkSession: SparkSession,
+                       paths: Seq[String],
+                       userSpecifiedSchema: Option[StructType],
+                       partitionSchema: StructType): PartitioningAwareFileIndex = {
+
+    class PartitionedMetadataLogFileIndex(
+                                           sparkSession: SparkSession,
+                                           path: Path,
+                                           parameters: Map[String, String],
+                                           userSpecifiedSchema: Option[StructType],
+                                           override val partitionSchema: StructType)
+      extends MetadataLogFileIndex(sparkSession, path, parameters, userSpecifiedSchema)
+
+    class PartitionedInMemoryFileIndex(
+                                        sparkSession: SparkSession,
+                                        rootPathsSpecified: Seq[Path],
+                                        parameters: Map[String, String],
+                                        userSpecifiedSchema: Option[StructType],
+                                        fileStatusCache: FileStatusCache = NoopCache,
+                                        userSpecifiedPartitionSpec: Option[PartitionSpec] = None,
+                                        metadataOpsTimeNs: Option[Long] = None,
+                                        override val partitionSchema: StructType)
+      extends InMemoryFileIndex(
+        sparkSession,
+        rootPathsSpecified,
+        parameters,
+        userSpecifiedSchema,
+        fileStatusCache,
+        userSpecifiedPartitionSpec,
+        metadataOpsTimeNs)
+
+    def globPaths: Boolean = {
+      val entry = options.get(DataSource.GLOB_PATHS_KEY)
+      Option(entry).forall(_ == "true")
+    }
+
+    val caseSensitiveMap = options.asCaseSensitiveMap.asScala.toMap
+    val hadoopConf = sparkSession.sessionState.newHadoopConfWithOptions(caseSensitiveMap)
+    if (FileStreamSink.hasMetadata(paths, hadoopConf, sparkSession.sessionState.conf)) {
+      new PartitionedMetadataLogFileIndex(
+        sparkSession,
+        new Path(paths.head),
+        options.asScala.toMap,
+        userSpecifiedSchema,
+        partitionSchema = partitionSchema)
+    } else {
+      val rootPathsSpecified = DataSource.checkAndGlobPathIfNecessary(
+        paths,
+        hadoopConf,
+        checkEmptyGlobPath = true,
+        checkFilesExist = true,
+        enableGlobbing = globPaths)
+      val fileStatusCache = FileStatusCache.getOrCreate(sparkSession)
+
+      new PartitionedInMemoryFileIndex(
+        sparkSession,
+        rootPathsSpecified,
+        caseSensitiveMap,
+        userSpecifiedSchema,
+        fileStatusCache,
+        partitionSchema = partitionSchema)
+    }
   }
 }
