+++ paimon-spark/paimon-spark4-common/src/main/scala/org/apache/paimon/spark/data/Spark4ArrayData.scala
@@ -20,12 +20,20 @@ package org.apache.paimon.spark.data
 
 import org.apache.paimon.types.DataType
 
-import org.apache.spark.unsafe.types.VariantVal
+import org.apache.spark.unsafe.types.{GeographyVal, GeometryVal, VariantVal}
 
 class Spark4ArrayData(override val elementType: DataType) extends AbstractSparkArrayData {
 
   override def getVariant(ordinal: Int): VariantVal = {
     val v = paimonArray.getVariant(ordinal)
     new VariantVal(v.value(), v.metadata())
+  }
+
+  def getGeography(ordinal: Int): GeographyVal = {
+    throw new UnsupportedOperationException("GeographyVal is not supported")
+  }
+
+  def getGeometry(ordinal: Int): GeometryVal = {
+    throw new UnsupportedOperationException("GeographyVal is not supported")
   }
 }
