+++ paimon-spark/paimon-spark4-common/src/main/scala/org/apache/paimon/spark/data/Spark4InternalRow.scala
@@ -21,12 +21,20 @@ package org.apache.paimon.spark.data
 import org.apache.paimon.spark.AbstractSparkInternalRow
 import org.apache.paimon.types.RowType
 
-import org.apache.spark.unsafe.types.VariantVal
+import org.apache.spark.unsafe.types.{GeographyVal, GeometryVal, VariantVal}
 
 class Spark4InternalRow(rowType: RowType) extends AbstractSparkInternalRow(rowType) {
 
   override def getVariant(i: Int): VariantVal = {
     val v = row.getVariant(i)
     new VariantVal(v.value(), v.metadata())
+  }
+
+  def getGeography(ordinal: Int): GeographyVal = {
+    throw new UnsupportedOperationException("GeographyVal is not supported")
+  }
+
+  def getGeometry(ordinal: Int): GeometryVal = {
+    throw new UnsupportedOperationException("GeographyVal is not supported")
   }
 }
