+++ paimon-spark/paimon-spark4-common/src/main/scala/org/apache/spark/sql/paimon/shims/MinorVersionShim.scala
@@ -18,92 +18,26 @@
 
 package org.apache.spark.sql.paimon.shims
 
-import org.apache.hadoop.fs.Path
 import org.apache.spark.sql.SparkSession
-import org.apache.spark.sql.catalyst.expressions.{Attribute, Expression}
-import org.apache.spark.sql.catalyst.plans.logical.{CTERelationRef, LogicalPlan, MergeAction, MergeIntoTable}
-import org.apache.spark.sql.execution.datasources.{DataSource, FileStatusCache, InMemoryFileIndex, NoopCache, PartitionSpec, PartitioningAwareFileIndex}
-import org.apache.spark.sql.execution.streaming.runtime.MetadataLogFileIndex
-import org.apache.spark.sql.execution.streaming.sinks.FileStreamSink
+import org.apache.spark.sql.catalyst.expressions.Expression
+import org.apache.spark.sql.catalyst.plans.logical.MergeRows.Instruction
+import org.apache.spark.sql.execution.datasources._
 import org.apache.spark.sql.types.StructType
 import org.apache.spark.sql.util.CaseInsensitiveStringMap
 
-
-import org.apache.hadoop.fs.Path
-
-import java.util.{Map => JMap}
-
-import scala.collection.JavaConverters._
-
 object MinorVersionShim {
 
+  def createKeep(context: String, condition: Expression, output: Seq[Expression]): Instruction = {
+    null
+  }
 
   def createFileIndex(
-                       options: CaseInsensitiveStringMap,
-                       sparkSession: SparkSession,
-                       paths: Seq[String],
-                       userSpecifiedSchema: Option[StructType],
-                       partitionSchema: StructType): PartitioningAwareFileIndex ={
-
-
-    class PartitionedMetadataLogFileIndex(
-                                           sparkSession: SparkSession,
-                                           path: Path,
-                                           parameters: Map[String, String],
-                                           userSpecifiedSchema: Option[StructType],
-                                           override val partitionSchema: StructType)
-      extends MetadataLogFileIndex(sparkSession, path, parameters, userSpecifiedSchema)
-
-    class PartitionedInMemoryFileIndex(
-                                        sparkSession: SparkSession,
-                                        rootPathsSpecified: Seq[Path],
-                                        parameters: Map[String, String],
-                                        userSpecifiedSchema: Option[StructType],
-                                        fileStatusCache: FileStatusCache = NoopCache,
-                                        userSpecifiedPartitionSpec: Option[PartitionSpec] = None,
-                                        metadataOpsTimeNs: Option[Long] = None,
-                                        override val partitionSchema: StructType)
-      extends InMemoryFileIndex(
-        sparkSession,
-        rootPathsSpecified,
-        parameters,
-        userSpecifiedSchema,
-        fileStatusCache,
-        userSpecifiedPartitionSpec,
-        metadataOpsTimeNs)
-
-
-    def globPaths: Boolean = {
-      val entry = options.get(DataSource.GLOB_PATHS_KEY)
-      Option(entry).forall(_ == "true")
-    }
-
-    val caseSensitiveMap = options.asCaseSensitiveMap.asScala.toMap
-    val hadoopConf = sparkSession.sessionState.newHadoopConfWithOptions(caseSensitiveMap)
-    if (FileStreamSink.hasMetadata(paths, hadoopConf, sparkSession.sessionState.conf)) {
-      new PartitionedMetadataLogFileIndex(
-        sparkSession,
-        new Path(paths.head),
-        options.asScala.toMap,
-        userSpecifiedSchema,
-        partitionSchema = partitionSchema)
-    } else {
-      val rootPathsSpecified = DataSource.checkAndGlobPathIfNecessary(
-        paths,
-        hadoopConf,
-        checkEmptyGlobPath = true,
-        checkFilesExist = true,
-        enableGlobbing = globPaths)
-      val fileStatusCache = FileStatusCache.getOrCreate(sparkSession)
-
-      new PartitionedInMemoryFileIndex(
-        sparkSession,
-        rootPathsSpecified,
-        caseSensitiveMap,
-        userSpecifiedSchema,
-        fileStatusCache,
-        partitionSchema = partitionSchema)
-    }
+      options: CaseInsensitiveStringMap,
+      sparkSession: SparkSession,
+      paths: Seq[String],
+      userSpecifiedSchema: Option[StructType],
+      partitionSchema: StructType): PartitioningAwareFileIndex = {
+    null
   }
 
 }
