+++ paimon-spark/paimon-spark4-common/src/main/scala/org/apache/spark/sql/paimon/shims/Spark4Shim.scala
@@ -30,6 +30,7 @@ import org.apache.spark.sql.catalyst.expressions.{Attribute, Expression}
 import org.apache.spark.sql.catalyst.expressions.aggregate.AggregateExpression
 import org.apache.spark.sql.catalyst.parser.ParserInterface
 import org.apache.spark.sql.catalyst.plans.logical.{Aggregate, CTERelationRef, LogicalPlan, MergeAction, MergeIntoTable}
+import org.apache.spark.sql.catalyst.plans.logical.MergeRows.Instruction
 import org.apache.spark.sql.catalyst.rules.Rule
 import org.apache.spark.sql.catalyst.util.ArrayData
 import org.apache.spark.sql.connector.catalog.{CatalogV2Util, Identifier, Table, TableCatalog}
@@ -58,9 +59,9 @@ class Spark4Shim extends SparkShim {
   }
 
   override def createSparkInternalRowWithBlob(
-                                               rowType: RowType,
-                                               blobFieldIndex: Int,
-                                               blobAsDescriptor: Boolean): SparkInternalRow = {
+      rowType: RowType,
+      blobFieldIndex: Int,
+      blobAsDescriptor: Boolean): SparkInternalRow = {
     new Spark4InternalRowWithBlob(rowType, blobFieldIndex, blobAsDescriptor)
   }
 
@@ -69,42 +70,42 @@ class Spark4Shim extends SparkShim {
   }
 
   override def createTable(
-                            tableCatalog: TableCatalog,
-                            ident: Identifier,
-                            schema: StructType,
-                            partitions: Array[Transform],
-                            properties: JMap[String, String]): Table = {
+      tableCatalog: TableCatalog,
+      ident: Identifier,
+      schema: StructType,
+      partitions: Array[Transform],
+      properties: JMap[String, String]): Table = {
     val columns = CatalogV2Util.structTypeToV2Columns(schema)
     tableCatalog.createTable(ident, columns, partitions, properties)
   }
 
   override def createCTERelationRef(
-                                     cteId: Long,
-                                     resolved: Boolean,
-                                     output: Seq[Attribute],
-                                     isStreaming: Boolean): CTERelationRef = {
+      cteId: Long,
+      resolved: Boolean,
+      output: Seq[Attribute],
+      isStreaming: Boolean): CTERelationRef = {
     CTERelationRef(cteId, resolved, output.toSeq, isStreaming)
   }
 
   override def supportsHashAggregate(
-                                      aggregateBufferAttributes: Seq[Attribute],
-                                      groupingExpression: Seq[Expression]): Boolean = {
+      aggregateBufferAttributes: Seq[Attribute],
+      groupingExpression: Seq[Expression]): Boolean = {
     Aggregate.supportsHashAggregate(aggregateBufferAttributes.toSeq, groupingExpression.toSeq)
   }
 
   override def supportsObjectHashAggregate(
-                                            aggregateExpressions: Seq[AggregateExpression],
-                                            groupByExpressions: Seq[Expression]): Boolean =
+      aggregateExpressions: Seq[AggregateExpression],
+      groupByExpressions: Seq[Expression]): Boolean =
     Aggregate.supportsObjectHashAggregate(aggregateExpressions.toSeq, groupByExpressions.toSeq)
 
   override def createMergeIntoTable(
-                                     targetTable: LogicalPlan,
-                                     sourceTable: LogicalPlan,
-                                     mergeCondition: Expression,
-                                     matchedActions: Seq[MergeAction],
-                                     notMatchedActions: Seq[MergeAction],
-                                     notMatchedBySourceActions: Seq[MergeAction],
-                                     withSchemaEvolution: Boolean): MergeIntoTable = {
+      targetTable: LogicalPlan,
+      sourceTable: LogicalPlan,
+      mergeCondition: Expression,
+      matchedActions: Seq[MergeAction],
+      notMatchedActions: Seq[MergeAction],
+      notMatchedBySourceActions: Seq[MergeAction],
+      withSchemaEvolution: Boolean): MergeIntoTable = {
     MergeIntoTable(
       targetTable,
       sourceTable,
@@ -113,6 +114,13 @@ class Spark4Shim extends SparkShim {
       notMatchedActions,
       notMatchedBySourceActions,
       withSchemaEvolution)
+  }
+
+  override def createKeep(
+      context: String,
+      condition: Expression,
+      output: Seq[Expression]): Instruction = {
+    MinorVersionShim.createKeep(context, condition, output)
   }
 
   override def toPaimonVariant(o: Object): Variant = {
@@ -136,11 +144,11 @@ class Spark4Shim extends SparkShim {
   override def SparkVariantType(): org.apache.spark.sql.types.DataType = DataTypes.VariantType
 
   override def createFileIndex(
-                                options: CaseInsensitiveStringMap,
-                                sparkSession: SparkSession,
-                                paths: Seq[String],
-                                userSpecifiedSchema: Option[StructType],
-                                partitionSchema: StructType): PartitioningAwareFileIndex = {
+      options: CaseInsensitiveStringMap,
+      sparkSession: SparkSession,
+      paths: Seq[String],
+      userSpecifiedSchema: Option[StructType],
+      partitionSchema: StructType): PartitioningAwareFileIndex = {
     MinorVersionShim.createFileIndex(
       options,
       sparkSession,
@@ -149,4 +157,3 @@ class Spark4Shim extends SparkShim {
       partitionSchema)
   }
 }
-
